Add files via upload
Disclaimer..I'm not a coder at all so I had a lot of help with the coding from various Ai but the ideas and features are all things I came up with and told the Ai to implement them. This is a work in progress. Feel free to take the code and improve it if you like. It's a Linux Midi Usb Host that has a Webgui for manual connections and is also standalone if you want to connect devices automatically with. The first device plugged in to the usb ports becomes the source device and the second device plugged in becomes the destination device. In manual mode you can connect devices in any direction and multiple devices if you need or want to. There are still some things that aren't quite working properly or at least with the midi devices I have available to me they don't seem to work as I hoped so feel free to install this and test your devices. The things that are working, auto mode seems to work as it should both devices connect and disconnect fine and midi messages are passed but as I said I have limited available midi device to test so you're mileage may vary. I'm hoping that at least a few people will see this and try it out and give me some feed back on how it works for them.